재고차감 인터페이스 정의 기존 비관적락 방식 외에 nativeQuery 방식 구현체를 추가하기 위함
