+++ src/test/java/com/example/demo/core/stock/service/DecreaseStockPessimisticLockServiceTest.java
@@ -34,11 +34,11 @@ import static org.mockito.Mockito.when;
 
 @IntegrationTest
 @RequiredArgsConstructor
-@DisplayName("DecreaseStockService")
-class DecreaseStockServiceTest extends TestDataInsertSupport {
+@DisplayName("DecreaseStockPessimisticLockService")
+class DecreaseStockPessimisticLockServiceTest extends TestDataInsertSupport {
 
     private final StockRepository stockRepository;
-    private final DecreaseStockService decreaseStockService;
+    private final DecreaseStockPessimisticLockService sut;
 
     @MockitoBean
     SoldOutProductService soldOutProductService;
@@ -84,7 +84,7 @@ class DecreaseStockServiceTest extends TestDataInsertSupport {
                 @Test
                 @DisplayName("정상적으로 재고를 차감하고, 재고가 모두 소진된 상품은 SoldOut()을 호출한다.")
                 void it() {
-                    decreaseStockService.decrease(param);
+                    sut.decrease(param);
 
                     List<Stock> actual = jpaQueryFactory.selectFrom(QStock.stock)
                         .where(QStock.stock.productCode.in(List.of("A202307300130", "A202307300131")))
@@ -122,7 +122,7 @@ class DecreaseStockServiceTest extends TestDataInsertSupport {
                 @Test
                 @DisplayName("InvalidStockQuantityException을 던지고, 차감한 재고를 롤백한다.")
                 void it() {
-                    assertThatThrownBy(() -> decreaseStockService.decrease(param))
+                    assertThatThrownBy(() -> sut.decrease(param))
                         .isExactlyInstanceOf(InvalidStockQuantityException.class)
                         .extracting("businessErrorCode")
                         .isEqualTo(INVALID_STOCK_QUANTITY);
@@ -161,7 +161,7 @@ class DecreaseStockServiceTest extends TestDataInsertSupport {
                 @Test
                 @DisplayName("BusinessException을 던지고, 차감한 재고를 롤백한다.")
                 void it() {
-                    assertThatThrownBy(() -> decreaseStockService.decrease(param))
+                    assertThatThrownBy(() -> sut.decrease(param))
                         .isExactlyInstanceOf(InvalidStockQuantityException.class)
                         .extracting("businessErrorCode")
                         .isEqualTo(INVALID_STOCK_QUANTITY);
@@ -218,7 +218,7 @@ class DecreaseStockServiceTest extends TestDataInsertSupport {
                         executorService.submit(() -> {
                                 try {
                                     startLatch.await();
-                                    decreaseStockService.decrease(param);
+                                    sut.decrease(param);
                                     successCount.incrementAndGet();
                                 } catch (Exception e) {
                                     failureCount.incrementAndGet();
@@ -278,7 +278,7 @@ class DecreaseStockServiceTest extends TestDataInsertSupport {
                         executorService.submit(() -> {
                                 try {
                                     startLatch.await();
-                                    decreaseStockService.decrease(param);
+                                    sut.decrease(param);
                                     successCount.incrementAndGet();
                                 } catch (Exception e) {
                                     failureCount.incrementAndGet();
@@ -344,7 +344,7 @@ class DecreaseStockServiceTest extends TestDataInsertSupport {
                         executorService.submit(() -> {
                                 try {
                                     startLatch.await();
-                                    decreaseStockService.decrease(param);
+                                    sut.decrease(param);
                                     successCount.incrementAndGet();
                                 } catch (Exception e) {
                                     failureCount.incrementAndGet();
@@ -414,7 +414,7 @@ class DecreaseStockServiceTest extends TestDataInsertSupport {
                         executorService.submit(() -> {
                                 try {
                                     startLatch.await();
-                                    decreaseStockService.decrease(param);
+                                    sut.decrease(param);
                                     successCount.incrementAndGet();
                                 } catch (Exception e) {
                                     failureCount.incrementAndGet();
